requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.13.3
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.17
email-validator==2.2.0
python-dateutil==2.9.0
